Decoupling AdminServer from ServerProcessor dependency
Summary:
This is a necessary refactoring to remove the direct dependency from Admin API server code from `Server` and `ServerProcessor`. This unblocks the project that will run this admin server as a standalone instance or locally in tooling.

The goal is that even if there are parts that require the `Server`, these will be disabled in such standalone instances and the rest of the API will work fine. An example is the per-shard RocksDB stats in FB303, or FailureDetector-based states.

Reviewed By: MohamedBassem, sh00s

Differential Revision: D13045202

fbshipit-source-id: be0592620b9bd3f1c268b566bfb63429f958ab8f