ClientReadStream: fix missing applying authoritative status when starting to read
Summary:
Without this diff ClientReadStream doesn't apply authoritative status until something times out or fails or or is updated. This doesn't seem right: (1) if many nodes are authoritative empty, reading may get stuck in the beginning until something times out; not sure how long that delay actually is (is reader-reconnect-delay=10ms the relevant timeout?), (2) worse than that, if a node is UNDERREPLICATION in event log and appropriately sends E::REBUILDING in STARTED messages, ClientReadStream ignores that and keeps treating the node as FULLY_AUTHORITATIVE; that makes UNDERREPLICATION not work most of the time. This diff fixes it.

Originally ClientReadStream was applying authoritative status in sendStart(): D2637266. Then D3962137 removed it saying that "We actually need to call applyShardStatus() only if we are re-connecting". I'm not sure what you meant by that: that applying status on startup is not needed, or that there's some other mechanism to apply status on startup? The latter wouldn't be true afaict. The former would be kind of questionable, and certainly breaks things in that UNDERREPLICATION+E::REBUILDING situation (which probably used to be ok at the time of the diff though, because ClientReadStream used to react to E::REBUILDING).

Reviewed By: lucaspmelo

Differential Revision: D13135244

fbshipit-source-id: ab66ed2471bc575679759e9e062eb289463806ad