Fix payload size check for string payloads
Summary:
Payload size was not properly checked when using client API in which payload is passed as string and owned by logdevice client.
Added a unit test which verified that it failed before and passed after the changes.
Decided to go for an overload approach to avoid confusion when payload is passed via constructor vs via setter.

Reviewed By: al13n321

Differential Revision: D13319420

fbshipit-source-id: f45766a0a3c5ac81d93b6b08ba54b4b68126146b