checking for graylisted nodes in NodeAvailabilityChecker
Summary:
Last step is taking the graylisting into consideration. This will be done in the NodeAvailaibilityChecker which is used by all the copyset selectors to build the copyset. If the node is graylisted by the new graylisting mechanism, it will be considered `NOT_AVAILABLE`. That's how the old graylisting is currently working.

I've also added the places where we will reset the graylist. Reseting the graylists in the new mechanism happens in two situations:
1. When we can't find a copyset, we reset the whole graylist (that's for V1) will revisit that later.
2. When graylisting is disabled.
The old graylisting used to clear the whole graylist when we reach the threshold, but this one ejects the older ones (to be revisited as well).

It's unfortunate that the graylisting retry thing is currently handled in every CopysetSelector individually, I might want to refactor this later.

Reviewed By: tau0

Differential Revision: D13194728

fbshipit-source-id: 05a4de65ee3e430c22690e2536d7d10e775d154c