Hardening the integration tests even more
Summary: Some of the tests are flaky because we have a small number of nodes each with 5 workers. With this many workers, some doens't talk to each other, so graylisting doesn't find outliers. This diff mainly decreases the number of workers to 2. I stress ran each test 100 times, should never be flaky afterwards.

Reviewed By: tau0

Differential Revision: D13376033

fbshipit-source-id: ed1bb956328968d54753a743f79911044db69e26