Fix ClientReadStream getting stuck in rewind loop when a node is drained
Summary:
In D13135244 I added a code path in ClientReadStream::rewind() that can potentially schedule another rewind. This seemed fine because it would only schedule a rewind when a node's authoritative status changes, which shouldn't happen often. But turns out there's a race condition: if, between the time rewind is scheduled and the time it actually happens, authoritative status gets overridden by STARTED_Message, the rewind will undo the override and will schedule another rewind, and this can repeat indefinitely.

The exact scenario (copied from task):
 1. Sender's status is AUTHORITATIVE_EMPTY. A rewind is scheduled.
 2. Before immediate_rewind_timer_ fires, we get a STARTED with E::OK and change status to FULLY_AUTHORITATIVE.
 3. Rewind happens. It changes authoritative status back to AUTHORITATIVE_EMPTY and schedules another rewind because of that.

For that to happen, a message roundtrip needs to happen faster than a zero-timeout timer fires. Surprisingly, this actually happens in practice, and somewhat reliably.

This diff fixes it by being careful about when to apply authoritative status and when to allow rewinds:
 * In `rewind()`, apply shard statuses before sending START messages. If the new shard statuses triggered a rewind, cancel it since we're about to rewind anyway. This is similar to the logic that was already there for `scd_->applyScheduledChanges()`.
 * In `updateCurrentReadSet()`, apply shard statuses before sending START messages. If the new shard statuses triggered a rewind, don't send START messages because rewind will do it anyway. This is similar to the logic that was already there for `scd_->updateStorageShardsSet()`.

Reviewed By: runemaster

Differential Revision: D13388981

fbshipit-source-id: 2eb3c1f12538739e295f84f6ffc0a2546b1b80ba