Fix compatibility of STORE_Header::STICKY_COPYSET
Summary:
Before D12863764 storage nodes write CSI iff STICKY_COPYSET flag is set in STORE_Message. After the diff, STICKY_COPYSET flag in incoming messages is as good as ignored. So far so good. But the diff also stopped setting the flag in outgoing messages from recovery and rebuilding. So if a newer version sends a recovery/rebuilding STORE to an older version, CSI won't be written, and the copy will be effectively lost. Data loss during rolling upgrades!

This diff fixes it by bumping protocol version and settings STICKY_COPYSET flag according to the old rules when talking to an old version of the server.

Reviewed By: runemaster

Differential Revision: D13212816

fbshipit-source-id: 7d7d73c841f714013f89b0eb60b415c33bf585cb