Use MarkFlushed notification to unregisterMemTable and move flush token window.
Summary:
Earlier MemTableRep destructor was used to unregister memtable and move the
flush window. But if there are active iterators they pin the memtable in memory
and the destructor does not get called for a long time. This stalls donors
proceeding forward during rebuilding without WAL.

In this diff we depend on explicit notification from rocksdb that the MemTable
is flushed instead on destruction of memtablerep. We should get this
notification irrespective of whether the memtable is pinned because of
iterators.

Reviewed By: scsiguy

Differential Revision: D13348454

fbshipit-source-id: 0f977f4e3accf0aa3d50f8fb872da23bc33ef087