Grouping outliers by node's scope
Summary:
In global clusters, store latency will be legitimately higher than inter DC latencies. For that reason, We need to run outlier detection per region instead of globally. To satisfy our outlier threshold, we take take biggest outliers from each region in a round robin fashion.
The decision of which region to take first is arbitrary but needs to be stable across runs to avoid invalidating the potential_graylist frequently, so I settled on using the region label for that.

Reviewed By: tau0

Differential Revision: D13361839

fbshipit-source-id: e73b23ae033bcfae26d2ac3f5d9f56637202c8ed