Get rid of --max-nodes setting
Summary:
This is for removing --max-nodes option in logdevice, which is used in logdevice/server/FailureDetector.cpp to preallocate required vectors. Preallocated vectors mean that 1) node_id needs to be consecutive, and 2) adding a node_id larger than max-nodes leads to an error.
To address this issue, existing vectors such as gossip_list_, gossip_ts_, failover_list_, and nodes_, are grouped and converted into unordered_map. Also, suspect_matrix_, an unused 2D vector, is eliminated.
GOSSIP_Message::serialize and ::deserialize are modified to convey the unordered_map between two nodes. In order to minimize code changes, gossip_list_, gossli_ts_, and failover_list_ are still in vector within GOSSIP message, and node_id_list_ is added to describe a series of node_id associated with vectors.

For now, same as before, unordered_map is preallocated during initialization. As a further work, when to add/delete a node to/from unordered_map needs to be discussed. Also, more test cases need to be added.

Reviewed By: mcrnic

Differential Revision: D13320758

fbshipit-source-id: 9691acc5a6eb2da97056f32abc2c6c3970e03ba4