Fix forward compatibility in EpochMetaData
Summary:
EpochMetaData::deserialize() has some provisions for forward compatibility: if it sees unknown flags, it assumes that they're produced by a newer version of the code, and it allows trailing extra bytes in this case (expecting that these bytes are some unknown fields serialized by the newer version of the code and gated by the unknown flags). But there's a flaw: the unknown flags seen during deserialization are not cleared and are serialized again. This is bad if these flags gate some fields we don't know about: we write the flags but don't write the fields. Then the new version of the code rightfully gets confused by that.

I ran into this when rolling out D13288949 to a test cluster.

Reviewed By: runemaster

Differential Revision: D13463100

fbshipit-source-id: 6529b9eef1158baa961ae51dbb4ccc9505d5eddd