Tool to generate write-side LDBench params for given tier
Summary:
This tool pokes the given tier to figure out what some of the LDBench
parameters should be that would simulate its load. This open-sourceable base
version only supports some of the parameters for a write worker. It's
structured to be easily extended for the closed source version, which will get
all necessary parameters for both read and write load.

This also fixes client API tracer support for dataSize(), and adds support for
PAYLOAD_HASH_ONLY in AsyncReader.

This currently accounts for a bug in dataSize() that causes overflow. It's
fixed in 2.38, so that part can be removed as soon as 2.37 is gone. It also
currently truncates the payload size distributions, which is probably fine for
now, but I'm gonna look into it more later in T37085637.

(Note: this ignores all push blocking failures!)

Reviewed By: al13n321

Differential Revision: D13147949

fbshipit-source-id: f8c2c5307b2ee23607de485b645fa4ce503a9795