DRR IO scheduler with an implementation for the storage thread pool.
Summary:
This review is for an IO scheduler for the storage tasks. The scheduler design is general purpose but only plugged into the storage thread pool. There are a lot of FIXME's that I will work on. But I want to start getting feedback so if I am missing something big then I have time to work on it -- thanks for helping with the review!

Design:

DRRScheduler.h implements the scheduler. There is one scheduler queue per shard. The queue is initialized with predefined shares per principle. The principles are directly mapped from priorities in StorageTaskPriority. The shares are defined in the settings file. The internal queue is a MPMC queue -- I didn't want to allocate memory and I was afraid that an intrusive list will not pass muster because the ListHead will need to be embedded into the enqueued request. There is a global lock in the scheduler. In unit tests the scheduler does about 2M reqs-per-sec from 4 threads.

ExecStorageThread dequeues and processes tasks. There are multiple threads dequeueing. The Scheduler dequeues the task based on turn and deficit. In addition, CompactionFilter also enqueues a psuedo CompactionThrottleStorageTask. This is a way to get compaction on the scheduler and have it process only as much as its share.

The scheduler is only implemented for the 'SLOW' queue and only if the settings is enabled. I'd like to extend it to the FAST and FAST_STALLABLE queues as well in the future.

Tasks can be scheduled based on their size or just per-request scheduling. The selection can be made using settings. I plan to productize (test and deploy) the request based scheduling first.

DRRTest.cpp has a several unit tests that test the correctness of share allocation, e.g., byte-based, req-based, etc.

Justin Gibbs, this is a much cleaner and much updated diff. Hopefully easier to review :).

Reviewed By: scsiguy

Differential Revision: D10502764

fbshipit-source-id: 9161120a1d2af404c404cdd0c5452bdcd4f2b181