pass all synchronous errors to callbacks in ZookeeperClient and NodesConfigurationStore
Summary:
The Zookeeper C API draws the distinction between synchronous errors and async
ones. The former are errors the client library can detect before making an
RPC to the server (e.g., invalid znode path) and are surfaced as return values
from the `zoo_aXXX` functions. The latter are (mostly) errors returned from the
server and they are passed to the callbacks.

To mimic this semantic makes the code unnecessarily complex; other
implementations may not have such a distinction / find it useful.

This diff changes the error handling semantics so that callbacks are always
invoked, the status code / error is always passed to the callback, making it
closer to a `Future`-s based API.

Differential Revision: D13400521

fbshipit-source-id: f398fcdec67aeb1798e6ae34d852efa279cc2ac6