admin_address should not be in Node config
Summary:
This removes the admin_address from the Nodes config. This is not needed and adds unnecessary complexity. The nodes config is a service discovery where nodes can use it to find each other. The Admin/Management ports are not needed in this context as nodes do not need to contact each other on these ports.

This makes handling of the Admin server address driven by settings similar to how we do with the existing command port.

Reviewed By: runemaster

Differential Revision: D13121346

fbshipit-source-id: 97ecc36465b47a57f030b9797e4bd55b8343b4cd