Accept TLS connections on gossip port
Summary:
When a new/existing gossip connection crosses the ssl-boundary
defined in the config file, we terminate the existing connection
and attempt to establish a new encrypted connection. However,
the gossip socket port currently supports plaintext connections
only. This change introduces a new config parameter:
ssl-on-gossip-port, which when enabled forces the gossip port
to accept ssl connections only.

Steps to enabled ssl on gossip port:
1. Set send-to-gossip-port = false. This will direct all gossip
   messages via the ssl-data port.
2. Set ssl-on-gossip-port = true. This will force gossip port
   to only accept ssl connections.
3. Set send-to-gossip-port = true. This will direct all gossip
   messages via the gossip port.

Steps to disable ssl on gossip port:
1. Set send-to-gossip-port = false
2. Perform a rolling restart
3. Set ssl-on-gossip-port = false
4. Set send-to-gossip-port = true

Reviewed By: davidvigier

Differential Revision: D13473296

fbshipit-source-id: f938af36386e24873de3de8399bce747fac01183