Clear potential graylist on reset
Summary: When testing graylisting on staging cluster, we noticed that the same node resets the graylist multiple times without respecting the grace period. So it turned out we don't clear the potential outlier list which causes nodes to get graylisted again in the next graylisting iteration. This diff clears the potential graylist as well on graylist reset.

Reviewed By: tau0

Differential Revision: D13527806

fbshipit-source-id: 1a19269bea6e1e8127ae6dce7aef6dfa119d3969