RocksDBMemTable manual flush policy per shard
Summary:
This diff implements flush policy and the related infrastructure for a single
shard. The implementation introduces a scheduleWriteBufFlush interface to localLogStore
to flush write buffers cached by a logstore shard. The interface accepts a
single shard active memory flush trigger, and target memory usage, if the memory usage is beyond budget. It
also allows to control the maximum size of a single write buffers if there are
multiple write buffers in the shard. The return value provides active, flushing/immutable and pinned memory usage after flushing memory usage after execution of api. The return values help the
controller to take some high level decisions.

For RocksDBLocalLogStore, the implementation get memtable stats on invocation
and if the memory usage or memtable size is beyond limits a memtable flush is
issued.

For PartitionRocksDBStore, a single shard will contain multiple column families
and hence multiple memtables. The details regarding choice of memtable is
covered in PartitionedRocksDBStore::FlushEvaluator::pickPartitionToFlush
method. The implementation for scheduleWriteBufFlush first picks the partitions and
flushes the memtables for each of them. This will change once we have the
rocksdb api to invoke atomic flush.

Reviewed By: al13n321

Differential Revision: D12897611

fbshipit-source-id: baa9ef37f8fd2ad0f5cfb88c064cf9de39b679f8