Quickly exiting watchdog thread when system is shutting down.
Summary:
Shutdown waits for 2 minutes. In this case, the watch dog thread was busy
printing the bt's of all the worker threads and could not complete work in
time. This lead to timeout and consequently coredump.

In this diff we will skip checking for stalls when workers have stopped
accepting work. And, if we get to printing stage and processor is already
marked shutdown we will exit the loop immediately.

There is a small window where the getting the backtrace takes a long time and
we never saw the flag, but I think that is ok and we cannot do much in that
case.

Differential Revision: D13409179

fbshipit-source-id: 317ba5b3212d862bbb9c4623739eab01e769b78b