Set MockStorageTask threadtype to SLOW
Summary: In recent change D13111277; the task types were shuffled around, as a result WriteStorageTask and hence MockWriteStorage task got type DEFAULT; the unit test looks for tasks of type SLOW, which was the previous type for StorageTasks. These need to be aligned so that the entries created by tryPutWrite are in the quote quried by tryGetWriteBatch.

Reviewed By: al13n321

Differential Revision: D13565022

fbshipit-source-id: 6ce93c3f3126109337d2d27fb23abfbb6a209a7b