Removing tapping into priority queue bucket.
Summary:
A sender can tap into priority queue during sending a packet or when workers run backlog. This created complicated interactions in code with the priority queue. Instead of this Justin was of the opinion that we should have this transfer in one place in the code.

There was a another motivation to do this, which was handling max bandwidth for bursts. This is handled in a previous diff. Briefly, guaranteed bw and overflow from other workers is deposited into a traffic class bucket during traffic shaper run and deposit budget was reset at every interval. Lets, assume size of the bucket is infinite. Hence, if a traffic class was silent it would end up getting guaranteed quota of credits into its bucket every iteration. This leaves some headroom before we hit the max bandwidth for the iteration. Now if a burst came in, the sender could transfer only difference between max_bw limit and current level, even though there are more credits in the priority queue bucket and this traffic class actually was silent for a while. Because of this, we could never reach max bw limit for this bursty traffic. Solution, was to transfer from priority queue bucket to traffic class bucket during the traffic shaper run proactively instead of waiting for requests to arrive and then tapping into priority queue.

Now, if buckets are sized correctly, then after distributing the credits from priority queue bucket to other traffic class bucket should leave no credits in the priority queue bucket. Hence, it made no sense to have tapping from priority queue during background run or while sending the packet. This was the other reason why this diff is useful.

Reviewed By: scsiguy

Differential Revision: D9441396

fbshipit-source-id: 0994c149214537d9f3d31f91ec8d168508277b5b