Fix backward compatibility issue because of HASHMAP_SUPPORT_IN_GOSSIP
Summary: It was reported that nodes before protocol version 92 (HASHMAP_SUPPORT_IN_GOSSIP)  e.g. 2.37, 2.38 etc. show all nodes with protocol version >= 92 as DEAD in 'info gossip' while nodes with protocol version 92 treat everyone as ALIVE. In a node with protocol version >= 92, node_list is based on hashmap, meaning that node_id is not ordered. On the other hand, in a node with protocol version < 92, node_list is based on vector, and all entries are strictly ordered from node_id of 0 to max. In a compatibility layer of GOSSIZP_Message::serialize(), when sending a gossip from new node to old node, node_list in a new node(sender) needed to be sorted because an old node(receiver) expects that nodes are all sorted in a gossip message. This issue is fixed in this diff.

Reviewed By: mcrnic

Differential Revision: D13572886

fbshipit-source-id: 50b8669a5f56eb307d7a8c403e16ab0ada7f659d