FileBasedNodesConfigurationStore for integration tests and cluster utility
Summary:
A quick stab at the `FileBasedNodesConfigurationStore`
implementation. We need it for integration test, cluster utility and OSS testing.

Use `flock()` to ensure inter-process access to the
store is serialized. For each key there is a lock file for `flock` and
data file to store the value. Within a process the store is single-threaded but can
be extended to multi-threaded by adding in-memory per-key shared
locks.

Another behavior is that the store provision the key entry on-demand
on the first write access and treat the created empty file as version 0.

Reviewed By: shixiao

Differential Revision: D13459473

fbshipit-source-id: 64dd256722293b37cb46861438a7bd17dfe4635e