AdminServer doesn't need FailureDetector, fixes crash in StandaloneAdminServer
Summary:
This removes the dependency of AdminServer on FailureDetector, this comes after I integrated the AdminServer with ClusterState instead. ClusterState is fetched remotely so the AdminServer doesn't need to be part of gossip. A future diff will be sent to make sure that the ClusterState is periodically polled.

This also removes the unused ISOLATED state. We can add it back if ClusterState can carry that information, I kept the enum value reserved.

Note: Without this, admin server crashes on ldshell status2.

Reviewed By: ylectric

Differential Revision: D13635013

fbshipit-source-id: 131d2369fad40a20e8e5047da0028ef7db26b795