Fix and simplify RocksDBListener
Summary: The hacky logic in RocksDBListener didn't age well. It's now broken for metadata CF and prints "please investigate" warnings all the time. But turns out the hack is not even needed anymore because rocksdb::TableFileCreationInfo now has enough information (cf name and file creation reason). So this diff kills it with fire.

Reviewed By: gdrane

Differential Revision: D13616075

fbshipit-source-id: d03441a5c86e321d8b70de3cd91217c4a225add3