Always build common_test_util as static
Summary:
When running unit tests, got failure with:

Double registration of singletons of the same underlying type; check for multiple definitions of type folly::Singleton<folly::ThreadWheelTimekeeper>

This comes from trying to link the same symbol twice, where normally the static linker would resolve this. The common_test_util was always inteneded for static linking, so make it so.

Reviewed By: tau0

Differential Revision: D13654086

fbshipit-source-id: b8a8fbce71cd42c29f65a60d1a13d8edafdeddea