Get rid of --max-nodes setting (2)
Summary:
This is in addition to previous diff, D13320758.
This is for allowing number of nodes to exceed max-nodes configuration. If a node-id is not found in nodes_ map in onGossipReceived(), the node-id is dynamically added to nodes_ map. A rw lock is added to protect read and write of nodes_ map.
The configuration max-nodes still exists but its definition is changed as "number of preallocated nodes in the cluster" If the name confuses users, I will rename it.
As of now, there is no case of deleting a node-id from nodes_ map.

Reviewed By: mcrnic

Differential Revision: D13505671

fbshipit-source-id: 24a61c9877d20dc21114e492fe550ba9461b7e91