Fix wrong ERROR message in BO
Summary:
In the digest phase, if LNG is seen to have an invalid OffsetMap, it aborts the
part that tries to recompte BO for each digest entries and prints an error message.

This will happen if Digest runs on a cluster where BO is disabled.
This error is showing up in dragon.staging.frc (which does not use BO).

I'm simply removing the error message.

Let me know if I understand something wrong, but from reading the code the same
issue was happening prior to Fayez's changes and were introduced in D8040509.

Reviewed By: runemaster

Differential Revision: D13674002

fbshipit-source-id: 5ee5bc7e7b6cdfe7d27cc7136a394f010e4854b2