max_background_flushes set to 1
Summary:
max_background_flushes * num_shards = total number of high pri rocksdb threads.
It was errorneously set to 15 which could have created 225 threads where we
just wanted 15. Fixing this here and in configs.

In prod we do not depend on this value which is was fortunate. We depend on
rocksdb-num-bg-threads-hi which is set correctly to 15 for most of the clusters.
But, it is incorrect for bryce canyon clusters as it should be 17.

Reviewed By: scsiguy

Differential Revision: D13535989

fbshipit-source-id: a90e80318637a53350372ed706238d1586b47c87