Fix RebuildingTest.MiniRebuildingAlwaysNonRecoverable
Summary: Under stress, this test fails sometimes because it couldn't find a node in the rebuilding set which was supposed to be there. This is because we took down the sequencer before some nodes had a chance to write the SHARD_NEEDS_REBUILD message. Though there was a check for self initiated rebuilding before the sequencer is taken down, it wasn't waiting for all nodes but rather just a single event. This diff fixes it by waiting for all dirty nodes to appear in rebuilding set before proceeding further.

Reviewed By: gdrane

Differential Revision: D13706654

fbshipit-source-id: 99f5d2dbba29b22d9307124e4ef6ecc6c31a81e6