Make nodeset selector pick both writable and unwritable nodes, and make nodeset bigger if many are unwritable
Summary:
See the comment that starts with "There's a subtlety about".

This also takes care of Xi's concern that when disks in repair are considered unwritable we'll regenerate too many nodesets. The consistent hashing will prevent that: only nodes that had that disk in the nodeset will change nodeset. That's good enough, right?

Reviewed By: scsiguy

Differential Revision: D13230614

fbshipit-source-id: d0cc5856ae30c6e1889fb8a75099b4eca9845655