Fix shard expansion in AdminAPI's safety checker
Summary:
In the Admin API interface we promise that a `ShardID(node_id, -1)` references all shards of a node. The expansion on the server side didn't actually expand it. On the server-side, the case of shard_index=-1 is invalid. This diff expands the request into a list of ShardID instances in this case.

Bonus: Adds a couple of missing fields to the thrift interface, the implementation is in the following diff.

Reviewed By: ylectric

Differential Revision: D13711465

fbshipit-source-id: c53e551b3d254828e3a98cf81963c01ec9a5e895