Fix DataKey compatibility
Summary:
In T10357210 I'm changing DataKey format to remove wave from key for good. Old format is ['d', log_id, lsn, 0xeeeeeeee] (0xeeeeeeee used to be wave), new format is ['d', log_id, lsn]. Since D10117646 we write records in the new format. Iterators recognize both old and new format, but without this diff they used to not detect or handle the situation when both keys are present for the same LSN. This broke things: an amend in new format won't apply to a record in old format; moreover, it'll make the record invisible because the iterator will see it as a dangling amend and skip the LSN.

This diff fixes it by making the iterator aware of the two formats. If there's a dangling amend in new format, the iterator will step forward to look for the corresponding full record in old format, then merge the two if it's found. This diff also makes logsdb proactively convert the unpartitioned records on startup; otherwise metadata logs may stay in old format indefinitely since they have infinite retention.

The merging logic has some correctness flaws to keep it simple, but that should be ok because (a) the common case is handled right, (b) metadata/internal logs are not affected because they're converted eagerly, (c) all the old-format records will be gone after retention.

Reviewed By: gdrane

Differential Revision: D13728812

fbshipit-source-id: bb01aa68fe10d71c8a0df27e1a58e6b7e45a0fe4