Fix potential deadlocks caused by shared_ptr misuse in ldbench
Summary:
Ldbench Worker used to pass weak reference to itself to append/read/findtime/etc callbacks. The callback would lock() the reference and if it's not null do something to it. If the temporary lock()ed shared_ptr happens to be the last reference to Worker (i.e. if the other reference was destroyed while the callback was running), the Worker's and therefore Client's destructor will be called from Client's callback, i.e. from the Client's own worker thread. That would deadlock or crash.

This diff gets rid of enable_shared_from_this, and instead makes sure that Client is destroyed (and therefore stops calling callbacks) before Worker's subclass is destroyed.

There was also a similar problem that signal handler could be the one to destroy Client. Surely Client destruction is not signal-safe. This diff fixes it too.

Reviewed By: gdavidsson

Differential Revision: D13689413

fbshipit-source-id: adc60a5b311b81fdd23f117655dc9697cf53137b