Mini-rebuilding awareness in safety-checker
Summary:
Currently safety-checker will consider a shard to be FULLY_AUTHORITATIVE if it has time-range rebuilding. This _can_ be dangerous as if we have enough of shards in this state we can lose read availability or rebuilding stall if we took more shards down.

Since there is no easy way to resolve this, I'm setting the authoritative status to UNAVAILABLE in the context of safety checker if a shard has mini-rebuilding. This is stricter than necessary but leans towards safety than otherwise.

Reviewed By: scsiguy

Differential Revision: D13669656

fbshipit-source-id: 172bc0741d89d5cf494182cf44a26a8909fc5772