Make the NodeRecoversInGracePeriod test more robust
Summary:
Fixes the `NodeRecoversInGracePeriod` test's flakiness by stopping the client appends instead of removing the latency injection as the node doesn't fully recover during the test duration after removing the injection. So instead, we can make it fail the min samples to be considered outlier test and eventually get removed from the graceperiod before getting graylisted.

This also fixes a bug in the test's appender, when stop is called multiple time.

Reviewed By: mcrnic

Differential Revision: D13760779

fbshipit-source-id: b2600258d04f26c0b719cc4f2e2d42a1a9295ec0