Remove ObserveOnly mode for sequencer boycotting
Summary: This diff removes the `boycotts-observe-only` flag. It's confusing, counter-intuitive and complicates the logic. The nodes will be marked as BOYCOTTED but you'll still see sequencers running on the node, and it complicates the logic because each time you check if the node is boycotted or not, you have to check if it's observe only or not. I've found a bug in the sequencer activation flow because we weren't checking this flag and it was causing the `DisableObserveOnly` test to be flaky.

Reviewed By: tau0, Zepheus

Differential Revision: D13750253

fbshipit-source-id: fa1e59812261e347f24ab3ca96bd6c142cb85804