Make AdaptiveBoycottDuration integration test less flaky
Summary:
The Adaptive boycotting test was a bit flaky because some nodes boycott N2 again once the initial boycott expires. Also the boycott duration was too short that it expiries before getting gossiped to the whole cluster.

For this test, we just want to make sure that the boycott just got doubled (or more).

Reviewed By: mcrnic

Differential Revision: D13232619

fbshipit-source-id: acbfc7fc7f09406541673758f9aaa6c8f4be68d3