Exit replication checker quickly in e2e's
Summary:
This diff resolves following things:
1. Replication checker goes over all the logs in e2e. Reduce that number to
10000 logs. That should be enough to catch bugs.
2. Time spent reading per log was not bound and was set to 240h. If we hit too
many gaps or for some reason checking is slow, the reading for the log kept
going on almost infinitely. Set a smaller value so that we complete replication checker per log
faster.
3. Add a option to stop replication checking on seeing too many failures.
Currently replication checker faithfully runs till the end. In e2e it
would be nice to fail fast because you already have lot of data to debug.

Reviewed By: ylectric

Differential Revision: D13743822

fbshipit-source-id: fb35739422df99aba69977ccb75c9fccec118661