Register onclose callback when sending logs config API messages in GetLogInfoFromNodeRequest
Summary:
It is possible that the shared_state_ has a socket_callback_ but it is not active. this happens because the associated socket got closed, the onclose callback executed and then deregistered.
In that case, we still need to pass it to sendMessage so that it is registered with a new socket.
If we don't and the socket gets closed again, no callback is executed and the config is not invalidated.

Reviewed By: mcrnic

Differential Revision: D13759271

fbshipit-source-id: 8c55a5e73f5e56ed0955675fbae35d047a379344