Collect unrecognized properties in ZookeeperConfig parser and implement support for URI to point to zookeeper ensemble
Summary:
Build a map of properties that are unknown to the ZookeeperConfig parser to be used by zookeeper client factories.
Also add support for "zookeeper_uri" property to eventually replace the qurorum. the format is a scheme, followed by "://", followed by a value. the standard way of specifying the quorum will then be:

zookeeper_uri: "ip://ip1:port1,ip2:port2,ip3:port3"

where ip1, ip2 and ip3 are the addresses of the quorum.

Reviewed By: Zepheus

Differential Revision: D13664814

fbshipit-source-id: c2ba86779127e02dc0b9a3ec21246f4e8f58f8ac