Split build and test steps in CircleCI
To allow the documentation build task to run, regardless of a unit test
failure, have split the jobs for build and unittesting.
Contents of the _build directory is preserved between workspaces.
For the unit test jobs we do not re-fetch the submodules; because we do
not need them during the unittests; but some files of the LogDevice
source tree are required.

Test Plan:

Ran config through CircleCI and it passed.