Split build and test steps in CircleCI (#65)
Summary:
To allow the documentation build task to run, regardless of a unit test
failure, have split the jobs for build and unittesting.
Contents of the _build directory is preserved between workspaces.
For the unit test jobs we do not re-fetch the submodules; because we do
not need them during the unittests; but some files of the LogDevice
source tree are required.
Pull Request resolved: https://github.com/facebookincubator/LogDevice/pull/65

Reviewed By: mzbeck

Differential Revision: D13819089

Pulled By: calebmarchent

fbshipit-source-id: 2be535866b1d88312d6f40f9895d80a33af988c1