Pullup checksumming evbuffer to prevent 2x Network Tx when TLS is enabled
Summary:
Right now TLS and checksumming are mutually exclusive to prevent a CPU regression that was discovered in 2.37
The reason was the way we write ProtocolHeader and Payload as 2 segments into output evbuffer, which causes 2 packets to be sent on network *only* if TLS is enabled.

This diff tries to serialize the header and payload into a single evbuffer(by using a pullup), and then finally move this one segment into output evbuffer. This prevents the CPU regression, at the cost of an extra memory copy, which acc. to testing didn't turn out to be a big deal in terms of CPU utilization.

Net improvement in CPU utilization:
Before this diff a comparison was done by Sandip b/w 2.36 and 2.37 with TLS enabled.
He noticed a jump in CPU utilization(relative to baseline) from 24% to 41%
With this diff, I'm seeing a jump from 24.69% to 28.16% only! See Testing section for actual numbers.

Differential Revision: D13728891

fbshipit-source-id: 847708319d2e7198d9e1678074e4907eb6a8a00b