Allow callbacks in ConfigurationFetchRequest
Summary:
In this diff, I'm modifying the `ConfigurationFetchRequest` to be able to accept callbacks. This also refactors the CONFIG_CHANGED::onReceived to make it a bit cleaner.

The flow is as follows:

If a callback is passed:
- Build a CONFIG_FETCH message passing the request_id to the header to indicate interest in the response.
- Start a timeout timer in case the response never came back.
- Add the request to a new worker map that maps the request id to the corresponding request.
- When a CONFIG_CHANGED message is received with the type `UPDATE`, it will resume the request with fetched config.
- The request will then pass the config to the callback, and delete itself.

If a callback is not passed:
- Build a CONFIG_FETCH message without passing the request_id in the header, which means that we are not interested in the response.
- Immediately delete the request given that we are not waiting for a response.

The logic in the `ConfigurationFetchRequest` assumes that the node that we will fetch the config from is passed in the constructor. And it doesn't actually do any retry mechanism. I can do that in later diffs, but by now this overlaps a lot with David's GET_OBJECT stuff, so I'll schedule something with David first.

In the next diffs,
1. I'll add a new request type to fetch the nodes configuration.
2. I'll wrap this request with a new read-only `NodesConfigSource`.

Reviewed By: runemaster

Differential Revision: D13487341

fbshipit-source-id: 6f67c025242cc3684ad61c3f7a9ae2f1e6dfbfd7