Change CONFIG_{FETCH,CHANHED} to avoid using FixedSizeMessage
Summary:
The next 3 diffs extends the `CONFIG_FETCH` and its response `CONFIG_CHANGED` message to allow them to accept callbacks. Which means that a user can create a request `ConfigurationFetchRequest` passing to it a callback and get the config back passed to the callback. The `ConfigurationFetchRequest` will be wrapped with a NodesConfigSource API which will serve as 1. The api doing the initial request during client bootstraping. 2. The source for the NCMs running on the client.

To do that, I need to add new fields to those two messages (request id & status). It turned out to be a hard thing because:
1. `CONFIG_FETCH` uses `FixedSizeMessage`, which makes it hard to add stuff to the header without breaking compatibility. This diff changes the message to stop using `FixedSizeMessage` and serialize the header manually. It also switches the header to be packed to be consistent with other messages (this should be safe since there's only a single field in the header anyways and it's covered in the tests).
2. `CONFIG_CHANGED` writes the header using `writer.write(header)`, but the header is not packed. So serializing each field manually doesn't work. We had a discussion around that in the team's group chat and eventually decided to deprecate the old struct and create a new packed one. The old one will still be there for clients with lower protocol version and should be removed when the min supported protocol version is past the new version;

This diff doesn't add any new fields, it just applies the changes mentioned above and adds tests for backward compatibility.

Reviewed By: runemaster

Differential Revision: D13487343

fbshipit-source-id: 45fa406d5242fccb1e484b112cdc74834d686d2a